+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 -   Font class included in public API.
 -   Glyphmap with dynamic font demo.
+-   Glyphset internal class for use in managing textures for Glyphmap - improve Glyphmap refresh performance.
+
+### Changed
+
+-   Glyphmap uses Glyphset internally and in renderers.
 
 ### Removed